btc-daemon-mn: update to mn 1.3.6, fix native-image
* Upgrade to Micronaut 1.3.6
* Add getters to BlockInfo.java
* Improve JsonRpcError handling for server
* Add more TypeHints for native-image
* Other tweaks/fixes

+++ bitcoinj-json/src/main/java/com/msgilligan/bitcoinj/rpcserver/BitcoinJsonRpc.java
@@ -10,7 +10,7 @@ public interface BitcoinJsonRpc {
     Integer getblockcount();
 // TODO: Support Sha256Hash type in request params?
 //    Object getblock(Sha256Hash blockHash, Integer verbosity);
-    Object getblock(String blockHashString, Long verbosity);
+    Object getblock(String blockHashString, Integer verbosity);
     Sha256Hash getblockhash(Integer blockNumber);
     Integer getconnectioncount();
     ServerInfo getinfo();
